First pass at taking advantage of key conversion
This still needs to be reviewed, but it’s a start at removing the
curve25519 private key that was part of the profile. Now that libsodium
supports key conversion, this isn’t needed.

A big upside will be to make key management easier.

I think this catches the key parts of the change, but make need some
editorial tweaks.